refactor: add card list component and home page
