+++ class-components/src/App.css
@@ -4,18 +4,3 @@
   padding: 2rem;
   text-align: center;
 }
-
-.header {
-  display: flex;
-  justify-content: center;
-  gap: 50px;
-  flex-wrap: wrap;
-}
-
-.cards {
-  margin-top: 50px;
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  gap: 20px;
-}
