+++ class-components/src/components/Card-list/Card-list.module.css
@@ -1,0 +1,7 @@
+.cardList {
+  margin-top: 50px;
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  gap: 20px;
+}
