+++ class-components/src/pages/Home/Home.module.css
@@ -1,0 +1,6 @@
+.header {
+  display: flex;
+  justify-content: center;
+  gap: 50px;
+  flex-wrap: wrap;
+}
